Data layout optimization
SYMGS and SPMV got significantly faster, doubling the performance of the implementation

Buffers for the transposed matrices are now created.

SYMGS still does not achieve maximum bandwidth, as applying the permutations in the kernel destroys the possibility of achieving coalesced memory accesses.

NEXT:
We can reorder all the matrices and vectors in the next step in order to get coalesced accesses and also remove accesses to the permutation matrix