Major performance boost from different coloring order
Change in how we generate the permutation matrix from the current colors, can reduce the extra iterations!
From 64 to 54 iterations. This change gives 11.5GFlops on a 1050, expected to give 80-85 on a Tesla V100, which is close to 50-60 percent of Nvidias implementation! ~70% of bandwidth.

Further minor improvements are:
- Small changes in SYMGS kernel, removed redundant operations on zeros on the first forward sweep of each iteration
- Removed the need of the diagonal vector to be used in the SyMGS kernel

These 2 minor changes provided a ~8-10% improvement on the 1050, but the impact on the Tesla V100 can be different.

The implementation now uses 90/128(70%) of the available Bandwidth