Tunable workgroups,data layout and CPU or GPU choice
The users can choose to use workgroup sizes of their choice. Furthemore they can choose to transpose the 2d arrays. This is usually better on GPUs.
A definition for CPU or GPU is also provided, which initializes the queue with cpu or gpu selector, as well as set more reasonable workgroup sizes(Can be
manualy tuned as well).